+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/Chore.kt
@@ -5,11 +5,12 @@ import com.altmann.choresmanager.models.Priority
 import kotlinx.datetime.DateTimePeriod
 import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalTime
 
 open class Chore(
     val choreId: Int,
-    val startTime: DateTimePeriod,
-    val endTime: DateTimePeriod,
+    val startTime: LocalTime,
+    val endTime: LocalTime,
     val daysOfWeek: List<DayOfWeek>,
     val startDate: LocalDate,
     val endDate: LocalDate,
@@ -24,7 +25,7 @@ open class Chore(
     val color: Color? = Color.DarkGray
 ) {
     open fun duration() : Int{
-        return endTime.hours - startTime.hours
+        return endTime.compareTo(startTime)
     }
 }