+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/grocery/GroceryChore.kt
@@ -6,11 +6,12 @@ import com.altmann.choresmanager.models.chores.Chore
 import kotlinx.datetime.DateTimePeriod
 import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalTime
 
 class GroceryChore(
     choreId: Int,
-    startTime: DateTimePeriod,
-    endTime: DateTimePeriod,
+    startTime: LocalTime,
+    endTime: LocalTime,
     daysOfWeek: List<DayOfWeek>,
     startDate: LocalDate,
     endDate: LocalDate,