+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/components/AddChorePopup.kt
@@ -1,6 +1,5 @@
 package com.altmann.choresmanager.ui.screens.components
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
@@ -20,19 +19,19 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.SubcomposeLayout
 import androidx.compose.ui.layout.onPlaced
-import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Popup
 import androidx.compose.ui.window.PopupProperties
 import com.altmann.choresmanager.models.Priority
 import com.altmann.choresmanager.models.chores.Chore
-import kotlinx.coroutines.flow.MutableStateFlow
+import com.altmann.choresmanager.utils.TimerParser
 import kotlinx.datetime.DatePeriod
-import kotlinx.datetime.DateTimePeriod
+import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalTime
 import kotlinx.datetime.plus
 
 @Composable
@@ -51,7 +50,7 @@ fun AddChorePopup(
         properties = PopupProperties(focusable = true),
     )
     {
-        popUpContent(
+        PopUpContent(
             popupOffset = popupOffset,
             datePicked = datePicked,
             date = date,
@@ -61,15 +60,19 @@ fun AddChorePopup(
 }
 
 @Composable
-fun popUpContent(
+fun PopUpContent(
     popupOffset: MutableState<Int>,
     datePicked: MutableState<String>,
     date: LocalDate,
     addChore: (chore: Chore) -> Unit,
 ) {
     val title = remember { mutableStateOf("") }
+    val startTimeTxt = remember { mutableStateOf(TextFieldValue("")) }
+    val endTimeTxt = remember { mutableStateOf(TextFieldValue("")) }
+    var selectedDays by remember { mutableStateOf<List<DayOfWeek>>(emptyList()) }
 
-    val datePickerVisible = remember { mutableStateOf(false) }
+    val startTime = remember {mutableStateOf(LocalTime(12, 0))}
+    val endTime = remember {mutableStateOf(LocalTime(14, 0))}
     Surface(
         shape = RoundedCornerShape(8.dp),
         tonalElevation = 4.dp,
@@ -85,23 +88,50 @@ fun popUpContent(
                 label = { Text("Chore Title") },
                 modifier = Modifier.padding(8.dp)
             )
-            Spacer(modifier = Modifier.height(4.dp))
+            FieldSpacer()
             TextField(
                 value = datePicked.value,
                 onValueChange = { datePicked.value = it },
                 maxLines = 1,
                 label = { Text("Due Date") },
-                modifier = Modifier.padding(8.dp)
-                    .clickable { datePickerVisible.value = true },
+                modifier = Modifier.padding(8.dp),
+            )
+            FieldSpacer()
+            TimeTextField(
+                timeTxt = startTimeTxt,
+                time = startTime,
+                label = "Start Time",
+                modifier = Modifier
+            )
+            FieldSpacer()
+            TimeTextField(
+                timeTxt = endTimeTxt,
+                time = endTime,
+                label = "End Time",
+                modifier = Modifier
+            )
+            Text(
+                "Days of Week",
+                style = MaterialTheme.typography.bodySmall,
+                modifier = Modifier.padding(start = 8.dp)
+            )
+            DayOfWeekPicker(
+                onSelect = { day ->
+                    selectedDays = if (selectedDays.contains(day)) {
+                        selectedDays - day
+                    } else {
+                        selectedDays + day
+                    }
+                }, selectedDays = selectedDays
             )
             Button(
                 onClick = {
                     addChore(
                         Chore(
                             choreId = 1,
-                            startTime = DateTimePeriod(hours = 10),
-                            endTime = DateTimePeriod(hours = 12),
-                            daysOfWeek = listOf(date.dayOfWeek),
+                            startTime = startTime.value,
+                            endTime = endTime.value,
+                            daysOfWeek = selectedDays,
                             startDate = date,
                             endDate = date + DatePeriod(days = 30),
                             choreException = listOf(),
@@ -121,4 +151,9 @@ fun popUpContent(
             }
         }
     }
+}
+
+@Composable
+fun FieldSpacer() {
+    Spacer(modifier = Modifier.height(4.dp))
 }